requirements.txt
# === Core Analysis Engine ===
astroid>=2.15.0,<3.0.0
pylint>=2.17.0,<3.0.0
pathspec>=0.11.0,<1.0.0
toml>=0.10.0,<1.0.0

# === Security & Quality Tools ===
semgrep>=1.45.0,<2.0.0
pip-audit>=2.6.0,<3.0.0
bandit>=1.7.0,<2.0.0
safety>=2.3.0,<3.0.0
detect-secrets>=1.4.0,<2.0.0
ruff>=0.1.0,<1.0.0
mypy>=1.5.0,<2.0.0
radon>=5.1.0,<6.0.0
black>=23.7.0,<24.0.0

# === Configuration & Data ===
pyyaml>=6.0,<7.0
dataclasses-json>=0.5.9,<1.0.0
typing-extensions>=4.7.0,<5.0.0

# === Performance & Testing ===
pytest>=7.4.0,<8.0.0
pytest-benchmark>=4.0.0,<5.0.0
pytest-timeout>=2.1.0,<3.0.0
pytest-json-report>=1.5.0,<2.0.0
memory-profiler>=0.61.0,<1.0.0
psutil>=5.9.0,<6.0.0

# === HTTP & Network (for integrations) ===
requests>=2.31.0,<3.0.0
urllib3>=2.0.0,<3.0.0

# === JSON & Data Processing ===
jsonschema>=4.19.0,<5.0.0